+++ data.rb
@@ -1,0 +1,20 @@
+$foods = [
+    { item: "Hotdog", price: 10.00 },
+    { item: "Sandwich", price: 15.00 },
+    { item: "Fish and Chips", price: 12.00 },
+    { item: "French Fries", price: 7.00 }
+]
+
+$drinks = [
+    { item: "Orange Juice", price: 5.00 },
+    { item: "Strawberry Juice", price: 5.00 },
+    { item: "Watermelon Juice", price: 5.00 },
+    { item: "Cola", price: 6.00 },
+    { item: "Water", price: 2.50 },
+]
+
+$candys = [
+    { item: "Cake", price: 8.00 },
+    { item: "Pudim", price: 5.00 },
+    { item: "Ice cream", price: 4.00 },
+]
